store: move stuff around
